+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.6.2
+
+### Breaking Changes
+
+None
+
+### Changed
+
+-   Added method `addPrivateKeys(privKeyA: PrivateKey, privKeyB: PrivateKey): PrivateKey` which takes two private keys and adds them together to return a new private key.
+-   Added method `subtractPrivateKeys(privKeyA: PrivateKey, privKeyB: PrivateKey): PrivateKey` which takes two private keys and subtracts `privKeyB` from `privKeyA` to return a new private key.
+
 ## 0.6.1
 
 ### Breaking Changes